website/requirements.txt
asgiref==3.2.10
Django==3.1.4
django-cors-headers==3.6.0
django-storages==1.10.1
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1
